% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjset_support.R
\name{adjset_support}
\alias{adjset_support}
\title{Compute support of adjustment set}
\usage{
adjset_support(
  dag_supp,
  x,
  y = 0L,
  adjset = "pa",
  dmats = NULL,
  replace_large_adjset = NULL
)
}
\arguments{
\item{dag_supp}{(list) posterior support over DAGs.}

\item{x, y}{(integer vector) column position(s) of cause and effect variable(s), respectively}

\item{adjset}{(character) name of adjustment set. See \code{\link[adjset:adjset]{adjset::find_adjset()}}.}

\item{dmats}{(optional) list of matrices representing the ancestral relationships
in each dag in \code{pdags[[1]]}. If not specified, \code{\link[=descendants]{descendants()}} is called for each
dag.}
}
\description{
Identify the set of unique adjustment sets in a sample of DAGs.
}
\examples{

data(bida_example_cat)
dags <- partitionMCMC$traceadd$incidence[-c(1:200)]

x <- match("X", colnames(dag))
y <- match("Y", colnames(dag))
adjset_support(dag_support(dags), x, y)

\dontrun{
library(adjset)
adjset_support(dag_support(dags), x, y, "o")
}
}
