% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjsets_support_from_cpdag.R,
%   R/adjsets_support_from_dags.R, R/parent_support_from_dags.R
\name{adjsets_support_from_cpdag}
\alias{adjsets_support_from_cpdag}
\alias{adjsets_support_from_dags}
\alias{parent_support_from_dags}
\title{List unique adjustment sets in a CPDAG}
\usage{
adjsets_support_from_cpdag(adjsets, cpdag)

adjsets_support_from_dags(
  adjsets,
  dags,
  support = rep(1/length(dags), length(dags)),
  checksize = NULL
)

parent_support_from_dags(dags, w = NULL, x = NULL, checksize = NULL)
}
\arguments{
\item{adjsets}{(character vector)
which adjset to identify. A subset of \code{c("anc", "pa", "pa_min", "o", "o_min")}.}

\item{cpdag}{(matrix)}

\item{dags}{(list)
a list of dags, e.g. a sample obtained with \link[BiDAG:partitionMCMC]{BiDAG::partitionMCMC}}

\item{support}{(numeric vector)
support for each dag in \code{dags}.}

\item{checksize}{(function, optional)
to evaluate if an adjustment set it to large. Defaults to \code{NULL}. See details.}

\item{maxsize}{(integer) maximum size of parent sets.
The support of parent sets \code{z} such that \code{length(z) > maxsize} is sat to zero.
If no parent set is smaller than \code{maxsize}, the empty set is returned as the unique parent set..}
}
\value{
\itemize{
\item \code{adjsets_support_from_dags} returns a list that includes for each adjset in \code{adjsets}:
\itemize{
\item \code{sets}: a \code{n-by-n} list-matrix, where \code{sets[[x, y]]} is a matrix
listing the unique adjsets w.r.t a cause \code{x} and effect \code{y} rowwise
\item \code{support}: a \code{n-by-n} list-matrix where \code{support[[x, y]]} is support over
the corresponding sets
}
}

\itemize{
\item \code{parent_support_from_dags}returns a \code{n} length list with elements:
\itemize{
\item \code{sets}: a matrix with the unique parents sets of each variable.
\item \code{support}: the support over parent sets.
}
}
}
\description{
Identify and compute support over unique adjustment sets in a list of DAGs.
}
\details{
\itemize{
\item \code{adjsets_support_from_cpdags}:
Identifies unique adjustment sets (parent or o-set only) locally in a CPDAG,
by directing siblings of every node. The same set of variables can consititute
the o-set in multiple of the resulting (P)DAGs, reflected in the returned "support".
}

\itemize{
\item \code{adjsets_support_from_dags}:
Computes support over unique adjustment set for each cause-effect pair (\code{x}, \code{y}),
where the adjustment eqauls \code{y} in all DAGs where \code{y} is not a descendant of \code{x}.
See also \link{adjsets_from_dag}.
\item \code{parent_support_from_dags}:
Computes support over unique parent sets of each variable, without encoding
zero-effects for non-descendants of the variables.
}
}
\seealso{
\link{adjsets_from_dag}
}
