% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor_mean.R
\name{backdoor_mean}
\alias{backdoor_mean}
\title{Compute the mean of an intervention distribution}
\usage{
backdoor_mean(x, nlevx)
}
\arguments{
\item{x}{object of class \code{\link{bida_bdeu}}}
}
\value{
the mean intervention distribution defined by \code{x}.
\itemize{
\item If \code{class(x) = "bida_bdeu"} this is a conditional probability table.
}
}
\description{
Compute the mean of an intervention distribution by applying the backdoor formula.
}
