% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor_params.R
\name{backdoor_params}
\alias{backdoor_params}
\title{Compute parameters for direct backdoor estimation of an intervention distribution}
\usage{
backdoor_params(type, data, x, y, z, hyperpar, lookup)
}
\arguments{
\item{data}{a N-by-n data matrix}

\item{x}{(integer)
column position of cause variable}

\item{y}{(integer)
column position of effect variable}

\item{z}{(integer vector) column position of the adjustment variables.
If \code{any(y == z)}, the marginal distribution of \code{y} is returned.
If \code{length(z) = 0}, the conditional distribution of \code{y} given \code{x}.}

\item{hyperpar}{list of parameters, depending on the prior distribution. See details.}
}
\value{
\itemize{
\item If \code{type} is in \code{c("cat", "ldag", "tree")}, an object of class \link{bida_bdeu}.
}
}
\description{
Compute parameters for direct backdoor estimation of an intervention distribution
}
