% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor_params_cat.R
\name{backdoor_params_cat}
\alias{backdoor_params_cat}
\alias{posterior_mean.backdoor_params_cat}
\alias{posterior_sample.backdoor_params_cat}
\title{Categorical intervention distributions}
\usage{
backdoor_params_cat(data, x, y, z, nlev, min_sparse = 2^10)

\method{posterior_mean}{backdoor_params_cat}(nyxz, ess, kx)

\method{posterior_sample}{backdoor_params_cat}(nyxz, samplesize, ess, kx)
}
\arguments{
\item{data}{a N-by-n matrix with integers representing a set of categorical variables.
Each variable is assumed to be coded as 0, 1, ..., nlev-1.}

\item{x}{(integer)
column position of cause variable}

\item{y}{(integer)
column position of effect variable}

\item{z}{(integer vector)
column position(s) of variable(s) in the backdoor set. A numeric/integer vector of
length 0 indicate an empty adjustment set. If the adjustment set includes \code{y},
\code{x} is assumed to have no effect on \code{y} and the marginal counts of \code{y} is returned.}

\item{nlev}{(integer vector)
vector of length n with the cardinality of each variable.}

\item{nyxz}{counts}

\item{ess}{(integer)
equivalent sample size}

\item{kx}{(integer)
number of possible interventions on the cause variable,
Used to replicate the marginal distribution when the causal effect is zero
(when \code{isZeroeffect(x) == TRUE}).
Otherwise the dimensions are given by \link{dim(x)} and this parameter is ignored.}

\item{samplesize}{(integer)
sample size}

\item{sparse}{(integer)
if the number of possible combinations exceeds this value, \code{prod(dims[c(y, x, z)]) > sparse},
the sample counts are stored in a sparse structure.
Otherwise, the counts are stored in an standard array.}
}
\value{
\itemize{
\item \code{backdoor_params_cat}: an array or a \link{bida_sparse_array} with sample counts.
\item \code{posterior_mean.backdoor_params_cat}: a kx-by-ky matrix with posterior means
\item \code{posterior_sample.backdoor_params_cat}: a kx-by-ky-by-\code{samplesize} array with interventional CPTs
}

\itemize{
\item \code{posterior_mean} a matrix with the posterior means of each intervention distribution,
i.e. a conditional probability table where element (x, y) equals is $P(Y = y| do(X = x))$
}

\itemize{
\item \code{posterior_sample} a 3-dimensional array with samples of the intervention
distribution parameters drawn from the posterior
}
}
\description{
Compute sample counts defining discrete intervention distributions
}
\details{
\code{backdoor_params_cat} computes sample counts for each configuration \verb{(y, x, z)}.
If \code{z} contains \code{y} the object contains the marginal counts of \code{y} (zero effect).
The dimension of the array is then of length 1.
Otherwise, over the triplet \verb{(y, x, z)}. Sufficient stats for the posterior
distribution over categorical intervention distributions.

\code{posterior_mean.backdoor_params_cat} computes (analytically) the posterior
mean values of the interventional CPT P(y|do(x))

\code{posterior_sample.backdoor_params_cat} samples a set of interventional CPTs
from the exact posterior, by sampling the relevant observational distributions
from their respective dirichlet posteriors.
}
