% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor_sample.R
\name{backdoor_sample}
\alias{backdoor_sample}
\title{Compute the mean of an intervention distribution}
\usage{
backdoor_sample(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{bida_bdeu}}}

\item{...}{additional arguments}
}
\value{
the posterior mean(s) of the distribution(s) specified by \code{x}
}
\description{
Compute the mean of an intervention distribution by applying the backdoor formula.
}
