% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdeu_posterior.R
\name{bdeu_posterior}
\alias{bdeu_posterior}
\title{Update BDeu-prior over a conditional probability table (CPT)}
\usage{
bdeu_posterior(data, y, x, ess, nlev, dx = NULL, sparse = TRUE)
}
\arguments{
\item{data}{(integer matrix) data over a set of categorical variables}

\item{y}{(integer) column position of outcome variable}

\item{x}{(integer vector) column position(s) of conditioning variable(s)}

\item{ess}{(numeric) equivalent sample size}

\item{nlev}{(integer vector) cardinality of the variables in \code{data}}

\item{dx}{(integer vector) joint outcome of \code{x} in data,
enumerated from 0 to \code{prod(nlev[x])}-1.
Included for settings where these joint outcomes can be reused.
If \code{NULL} the function computes the joint outcomes as needed (default).}

\item{sparse}{(logical) If \code{TRUE}, the counts are stored in a \link{bida_sparse_array}
object (default). Otherwise, the counts are stored in a standard array.
Otherwise, the function returns an object of class \link[base:array]{base::array}}

\item{`ayx`}{an array of}
}
\value{
\itemize{
\item \code{bdeu_posterior}: an \link{bida_sparse_array} or \link[base:array]{base::array} with dimensions
\code{nlev[c(y, x)]}, with parameters of the posterior Dirichlet distributions.
}
}
\description{
Update BDeu-prior over a conditional probability table (CPT)
}
