% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_.R
\name{bida}
\alias{bida}
\alias{posterior_mean.bida}
\alias{posterior_sample.bida}
\title{Class: bida}
\usage{
bida(ps, data, type, params, pairs = NULL)

\method{posterior_mean}{bida}(x, ...)

\method{posterior_sample}{bida}(x, n, ...)
}
\arguments{
\item{ps}{(list)
Support over adjustment sets, stored in a list with two named elements \code{sets} and \code{support}.
See \link{adjsets_support_from_dags} \link{parent_support_from_dags} \link{parent_support_exact}}

\item{data}{(matrix)}

\item{params}{additional params}

\item{pairs}{(matrix) a n-by-n matrix indicating which pairs should be considered.
If NULL (default), all pairs are considered and \code{pairs[i, j] = 1} for all but \code{i == j}.}

\item{x}{an object of class \link{bida}}

\item{samplesize}{(integer) samplesize}
}
\value{
\code{bida}: a n-by-n list where each element is an object of class \code{bida_pair}
\code{posterior_mean}: a n-by-n list with the posterior mean
\code{posterior_sample}: a n-by-n list with posterior samples
}
\description{
An object of class bida is a matrix list with objects of class \link{bida_pair},
that describes the posterior mixture distribution over the causal effect between every variable pair.
The \code{posterior_mean} and \code{posterior_sample} methods calls the respective methods
for each \link{bida_pair} and collect the output in a matrix list.
}
\seealso{
\link{bida_pair_cat}
}
