% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_exact.R
\name{bida_exact}
\alias{bida_exact}
\title{Main function for computing the BIDA posterior using dynamic programming.}
\usage{
bida_exact(
  data,
  type = "",
  max_parent_size = min(5, ncol(data) - 1),
  temp_id = "",
  filepath_apssolver = "./aps-0.9.1/aps",
  ess = 1,
  nlev,
  dirapprox = FALSE
)
}
\arguments{
\item{data}{an n-by-d data matrix}

\item{type}{data type ("Categorical" or "Gaussian")}

\item{max_parent_size}{max size of parent sets (default: 4)}

\item{temp_id}{name for file in which scores are temporarily saved}

\item{filepath_apssolver}{filepath directing to the apssolver}

\item{ess}{equivalent sample for BDeu prior (default: 1). Ignored if \link{type != "Categorical"}}

\item{dirapprox}{indicator for whether parameters of the approximate dirichlet posterior of the intervention probabilities
should be computed. If FALSE, the posterior parameters of the dirichlet distribution of the joint probability p(x, y, z) is returned.}
}
\value{
bida object containing the parameters that specify the
}
\description{
Main function for computing the BIDA posterior using dynamic programming.
}
