% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_pair_cat.R
\name{bida_pair_cat}
\alias{bida_pair_cat}
\title{Class: \code{bida_pair}}
\usage{
bida_pair_cat(data, x, y, sets, support, nlev, ess = 1)
}
\arguments{
\item{data}{a N-by-n data matrix}

\item{x}{(integer)
column position of cause variable}

\item{y}{(integer)
column position of effect variable}

\item{sets}{(integer matrix)
each row contains column positions of a valid adjustment set.
NA entries in each row is ignored.}

\item{support}{(numeric vector)
a vector with support for each adjustment set. Should be of length \code{nrow(sets)}
and the entries should sum to 1 (not checked).}

\item{nlev}{(integer vector)
cardinality of each variable.}

\item{ess}{(numeric)
imaginary sample size. Defaults to 1.}
}
\value{
An object of class \link{bida_pair_cat} is a list that contains:
\itemize{
\item \code{params}:
\item \code{support}:
\item \code{nlev}:
\item \code{ess}:
}
}
\description{
Compute parameters defining the posterior mixture distribution over the
parameters defining interventional distribution p(y|do(x)).
}
\examples{
nlev <- rep(2, 3)
lev  <- lapply(nlev-1, seq.int, from = 0)
data <- as.matrix(expand.grid(lev))
sets <- matrix(c(NA, 2), nrow = 3)
fit <- bida_pair_cat(data, 1, 2, sets, rep(1/3, 3), nlev, ess = 0)
fit
# compute posterior mean
posterior_mean(fit)
# sample from postrior
posterior_sample(fit)


}
