% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_posterior.R
\name{bida_posterior}
\alias{bida_posterior}
\title{Compute posterior intervention probability}
\usage{
bida_posterior(ps, data, type, hyperparams)
}
\arguments{
\item{ps}{nested list with parent_sets and parent score for each variable. See X.}

\item{data}{data frame with observered variables}

\item{nlevels}{integer vector. Number of categories of each variable}

\item{ess}{imaginary sample size for bdeu-prior}

\item{th}{threshold value for discarding parent set. Sorted by the}

\item{dirapprox}{bolean. If TRUE the posterior distribution is approximated as a Dirichlet.}
}
\value{
a list with parameters of the approximate posterior distribution of the
intervention probabilities.
}
\description{
Compute posterior intervention probability
}
\details{
\itemize{
\item \link{posterior_sample} return samples from the posterior mixture distribution,
First, adjustment sets from the posterior distribution \eqn{p(G_{ij}|D)} are sampled.
Given the adjustment set of each observation, one can sample observational
parameters from the corresponding (local) Dirichlet distributions.
The interventional distributions are then given by applying the back-door-formula.
If specified, \link{ace_funs} is applied to the sample of interventional CPTs,
in which case the function returns a list of matrices, one matrix for each
function in \link{ace_funs}, where element (i, j) contains the transformed CPTs.
}
}
