% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_posterior_cat.R
\name{bida_posterior_cat}
\alias{bida_posterior_cat}
\alias{bida_posterior_cat_local}
\alias{posterior_mean.bida_posterior_cat}
\alias{posterior_sample.bida_posterior_cat}
\title{Compute the BIDA posterior over categorical intervention distributions}
\usage{
bida_posterior_cat(ps, data, x, y, ess, nlev, Nys = NULL)

bida_posterior_cat_local(ps, data, x, ys, ess, nlev, Nys = NULL)

\method{posterior_mean}{bida_posterior_cat}(obj, contrasts = list())

\method{posterior_sample}{bida_posterior_cat}(obj, n, contrasts = list())
}
\arguments{
\item{ps}{(list) posterior support over adjustment sets.}

\item{data}{(integer matrix) matrix with categorical variables}

\item{x}{(integer) column position of cause variable}

\item{y}{(integer) column position of effect variable.}

\item{ess}{(numeric) imaginary sample size}

\item{nlev}{(integer vector) cardinality of the variables in \code{data}.}

\item{Nys}{(list) list with parameters for the posterior dirichlet
distribution over the marginal distribution of variable.
If \code{NULL} these parameters are computed as needed (default).
When the function is called for each cause-effect pair,
these can be pre-computed for speed and memory.}

\item{ys}{(integer vector) column position(s) of effect variable(s).}

\item{contrasts}{(list) a list of functions for contrasting IPTs and summarize
the effect of interventions into a single value, such as \code{\link[=jsd]{jsd()}}.
The functions are assumed to return 0 when the interventional distributions
are identical.}

\item{n}{sample size}
}
\value{
an object with class \code{posterior_bida} and
subclass \code{bida_posterior_cat}. The local version returns a
list of such objects.
}
\description{
Compute the approximate posterior distribution over the interventional probability
table (IPT) for a cause-effect pair \verb{(x, y)}, given the posterior over a set
of adjustment sets.
The local version \code{\link[=bida_posterior_cat_local]{bida_posterior_cat_local()}} can be used the parent sets of \code{x}
is used for adjustment. Allows for some speed gain by re-using the calculations
involving \code{x} and each adjustment set for each effect variable \code{y} in \code{ys}.
}
\section{Methods}{

\itemize{
\item \code{posterior_mean()} returns the posterior mean IPT or the posterior mean
causal effect if \code{contrasts} is specified.
}


\itemize{
\item \code{posterior_sample()} returns a posterior sample of
intervention probability tables (IPTs) or causal effects if \code{contrasts}
is specified.
}
}

\examples{

data(bida_example_cat)
attach(bida_example_cat)

# compute dag posterior from sample of DAGs
dags <- partitionMCMC$traceadd$incidence[-c(1:200)]
dag_supp <- dag_support(dags)

# compute adjset posterior
x <- match("X", colnames(dag))
y <- match("Y", colnames(dag))
ps <- adjset_support(dag_supp, x, y, "pa")

# compute bida posterior
pb <- bida_posterior_cat(ps, data, x, y, ess = 1, nlev = nlev)
pb

posterior_mean(pb)
posterior_sample(pb, 10)




\dontrun{
# profile
n    <- 20
N    <- 1000
dags <- replicate(100, rand_dag(n, 4), simplify = FALSE)
nlev <- rep(3, n)
data <- vapply(nlev, sample.int, size = N, replace = T, integer(N))-1
ess  <- 1

x <- 1
ps  <- adjset_support(dag_support(dags), x, 0L, "pa")
library(microbenchmark)
microbenchmark(bida_posterior_cat_local(ps, data, x, 1:n, ess, nlev),
               lapply(1:n, function(y) bida_posterior_cat(ps, data, x, y, ess, nlev)),
               times = 10, check = "equivalent")
}


}
