% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_sample.R
\name{bida_sample}
\alias{bida_sample}
\title{Main function for computing the BIDA posterior using a sample of DAGs.}
\usage{
bida_sample(
  data,
  dags,
  support = rep(1/length(dags), length(dags)),
  type = "categorical",
  adjsets = "pa",
  params = NULL
)
}
\arguments{
\item{data}{an n-by-d data matrix}

\item{dags}{list of DAGs (adjacency matrices, obtained e.g. with BiDAG::partitionMCMC)}

\item{support}{numeric vector of same length as \link{dags}. Support of each DAG. Default to \link{1/length(dags)}.}

\item{type}{data type ("Categorical" or "Gaussian")}

\item{adjset}{type of adjustment set to compute in each DAG.}
}
\value{
an object of class \link{bida}
}
\description{
Main function for computing the BIDA posterior using a sample of DAGs.
}
\seealso{
\link{bida}
}
