% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_from_cpt_arrays.R
\name{bn_from_cpt_arrays}
\alias{bn_from_cpt_arrays}
\title{Construct a bnlearn-network from a list of CPTs}
\usage{
bn_from_cpt_arrays(cpts)
}
\arguments{
\item{cpts}{a list of arrays representing the CPT of each node.}
}
\value{
An object of class \code{bn.fit.dnet}
}
\description{
Construct a bnlearn-network from a list of CPTs
}
\keyword{internal}
