% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_no_points.R
\name{boxplot_no_points}
\alias{boxplot_no_points}
\title{Boxplot with no outliers}
\usage{
boxplot_no_points(
  df,
  x = "N",
  y = "value",
  fill = NULL,
  color = NULL,
  group = NULL,
  color_mean = NULL,
  alpha = 1,
  width = 0.5
)
}
\arguments{
\item{df}{(data.frame)}

\item{x, y, fill, color, group}{(character) name of variables to corresponding arguments of \code{\link[ggplot2:aes]{ggplot2::aes()}}}

\item{color_mean}{(character)}

\item{alpha}{(numeric)}

\item{width}{(numeric)}
}
\value{
a ggplot2-object
}
\description{
Boxplot with no outliers
}
\keyword{internal}
