% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prec_recall.R
\name{compute_prec_recall}
\alias{compute_prec_recall}
\title{Compute precision recall}
\usage{
compute_prec_recall(x, y, method)
}
\arguments{
\item{x}{score}

\item{y}{outcome}

\item{method}{which method to use}
}
\value{
a list with the following elements:
\itemize{
\item \code{df}: a data frame with FPR, TPR and PPV at each evaluated value of \code{x}
\item \code{avgppv}: the average precision
}
}
\description{
Compute precision recall
}
\examples{

# no ties
compare_prec_recall(1:0, 1:0, niter = 10)
compare_prec_recall(0:1, 1:0, niter = 10)

# perfectly right
x <- c(2, 2, 1, 0)
y <- c(1, 1, 1, 0)
compare_prec_recall(x, y, niter = 20)

# perfectly wrong
x <- c(0, 0, 1, 1)
y <- c(1, 1, 0, 0)
compare_prec_recall(x, y, niter = 20)


#' # get first, miss last
x <- c(1, 1, 0)
y <- c(1, 0, 1)
compare_prec_recall(x, y, niter = 20)

# get first, miss last, no ties
x <- c(2, 1, 0)
y <- c(1, 0, 1)
compare_prec_recall(x, y, niter = 20)

x <- c(8:1, rep(0, 2))
y <- c(x[x > 0]/10 > runif(8), .2 > runif(2))
compare_prec_recall(x, y, niter = 20)

x <- c(80:1, rep(0, 20))
y <- c(x[x > 0]/100 > runif(80), .2 > runif(20))
compare_prec_recall(x, y, niter = 20)



}
\keyword{internal}
