% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_from_data_matrix.R
\name{counts_from_data_matrix}
\alias{counts_from_data_matrix}
\title{Fast cross tabulation of categorical data}
\usage{
counts_from_data_matrix(data, nlev, sparse = F)
}
\arguments{
\item{data}{(integer matrix)
each column represent observed values of a categorical variable.}

\item{nlev}{(integer vector)
number of categories of each variable.}

\item{sparse}{(boolean)
if TRUE, a \link{bida_sparse_array} with counts is returned.
If false, the counts are returned in a \link[base:array]{base::array} with dimensions given by \code{nlev}.
Default to false.}
}
\value{
An array with counts over each joint combination in \code{data}
}
\description{
Faster version of \code{table(as.data.frame(data))} for numeric matrices with
integer values.
}
\details{
The ith variable (column i of \code{data}) is assumed to take values from \code{0} to \code{nlev[i]-1}.
This is not checked.
}
\keyword{internal}
