% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpquery_from_cpt_arrays.R
\name{cpquery_from_cpt_arrays}
\alias{cpquery_from_cpt_arrays}
\title{Exact computation of conditional probability distributions}
\usage{
cpquery_from_cpt_arrays(cpts, y, x = integer(0), anc = NULL)
}
\arguments{
\item{cpts}{(list)
list of \link[base:array]{base::array}s that specifies a conditional probability table $P(X_i| Pa(X_i))$
For each array the \code{dimnames} attribute must be a named list, indicating the
scope of the vector - i.e. the names of X_i and Pa(X_i).}

\item{y}{(integer vector)
position of the (set of) variable(s) in \code{cpts}}

\item{x}{(integer vector)
position of variables that constitutes the conditioning set.
Defaults to the \code{integer(0)}}

\item{anc}{(logical vector)
indicator of ancestors of \code{y} and \code{x}.
If \code{NULL}, the ancestors is computed via the DAG implied by \code{cpts}.}
}
\value{
an array with the conditional distribution of \code{y} given \code{x}.
}
\description{
Compute conditional probability distributions from a set of CPTs associated
associated with a Bayesian network, using the variable elimination algorithm.
}
\keyword{internal}
