% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_support.R
\name{dag_support}
\alias{dag_support}
\title{Compute support of DAG posterior}
\usage{
dag_support(dags, p = rep(1/length(dags), length(dags)))
}
\arguments{
\item{dags}{list of adjacencey matrices}

\item{p}{vector of weights. Uniform by default.}
}
\value{
a list representing with the unique DAGs in \code{dags} and their relative
freuency
}
\description{
Compute support of DAG posterior
}
