% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{descendants}
\alias{descendants}
\alias{descendants.default}
\alias{descendants.graphNEL}
\alias{descendants.bn.fit}
\title{Find ancestors relations corresponding to a network}
\usage{
descendants(x)

\method{descendants}{default}(x)

\method{descendants}{graphNEL}(x)

\method{descendants}{bn.fit}(x)
}
\arguments{
\item{x}{a n-by-n adjacency matrix or a bn.fit object}
}
\value{
a n-by-n adjacency matrix for the acnestor relations
}
\description{
Find ancestors relations corresponding to a network
}
