% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_tex.R
\name{df_to_tex}
\alias{df_to_tex}
\title{Title}
\usage{
df_to_tex(
  df,
  values_from = "value",
  names_from = character(0),
  group_by = character(0),
  file = "",
  caption = NULL,
  label = NULL,
  digits = 2,
  group_keys_in_rows = FALSE,
  footnote = "",
  addTimeStamp = FALSE,
  bold_fun = NULL,
  sanitize.text.function = NULL
)
}
\arguments{
\item{sanitize.text.function}{}
}
\description{
Title
}
\examples{

df <- data.frame(group = rep(letters[1:3], each = 6),
                 key   = rep(rep(1:3, each = 2), 3),
                 name  = rep(rep(LETTERS[1:2], 3), 3),
                 value1 = 1:18,
                 value2 = 18:1)

values_from <- c("value1", "value2")
names_from  <- "name"
group_by    <- "group"
df_to_tex(df, values_from, names_from, group_by)

# put group keys in separating row
df_to_tex(df, values_from, names_from, group_by, group_keys_in_rows = TRUE)

# make maximum of each row bold
df_to_tex(df, values_from, names_from, group_by, group_keys_in_rows = TRUE, bold_fun = "max")

}
