% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_networks.R
\name{example_networks}
\alias{example_networks}
\title{Title}
\usage{
example_networks(name)
}
\arguments{
\item{name}{(character) the name of an example network}
}
\value{
a bn.fit object
}
\description{
Title
}
\examples{
bn <- example_networks("illustrate_posteiror")
g <- as
}
