% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_grid_fast}
\alias{expand_grid_fast}
\title{Faster version of expand.grid for integer values.}
\usage{
expand_grid_fast(x = NULL, k = NULL)
}
\arguments{
\item{x}{a list with integer vectors.}

\item{k}{the length of each vecotr in \code{x}.}
}
\value{
a \code{k-by-k} matrix with all combinations of the vectors in \link{x}.
In contrast \link{base}\link{expand.grid} returns a data.frame.
}
\description{
Faster version of expand.grid for integer values.
}
\examples{
x = list(0:1, 0:1, 0:1)
all(expand.grid(x) == bida:::expand_grid_fast(x))
microbenchmark::microbenchmark(as.matrix(expand.grid(x)), bida:::expand_grid_fast(x))

}
\keyword{internal}
