% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{combine_char_vectors_c}
\alias{combine_char_vectors_c}
\alias{named_array_c}
\alias{scope_c}
\alias{factor_product_c}
\alias{factors_product_c}
\title{Compute product of two factors}
\usage{
combine_char_vectors_c(x, y)

named_array_c(values, dim, names)

scope_c(x)

factor_product_c(x, y)

factors_product_c(factors)
}
\arguments{
\item{x, y}{numeric vectors with:
\itemize{
\item dimensions/cardinalities given by \code{x.attr("dim")} and similarly for \code{y}
\item scope given by names of \code{x.attr("dimnames")} and similarly for \code{y}
}}
}
\description{
Compute product of two factors
}
\seealso{
cpqueries_from_cpt_arrays
}
\keyword{internal}
