% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famscore_bdeu.R
\name{famscore_bdeu}
\alias{famscore_bdeu}
\title{Compute BDeu family score for categorical data}
\usage{
famscore_bdeu(m, ess = 1)
}
\arguments{
\item{m}{a \code{kx}-by-\code{kpa} matrix with counts (a frequency table).}

\item{ess}{equivalent sample size parameter.}
}
\value{
family score.
}
\description{
Compute BDeu family score for categorical data
}
