% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famscore_to_file_fml.R
\name{famscore_to_file_fml}
\alias{famscore_to_file_fml}
\title{Computes FML family score for Gaussian (zero-centered) data.}
\usage{
famscore_to_file_fml(data, max_parent_size = ncol(data) - 1, file_out)
}
\arguments{
\item{data}{an n-by-d data matrix (Gaussian zero-centered data).}

\item{max_parent_size}{limit on the maximum parent set size.}

\item{file_out}{name of the file to which the scores are written (".score"}
}
\value{
a file with the parent scores in GOBNILP format.
}
\description{
Computes FML family score for Gaussian (zero-centered) data.
}
