% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parts.R
\name{get_parts}
\alias{get_parts}
\title{Unlist a partition}
\usage{
get_parts(P, lens = lengths(P))
}
\arguments{
\item{P}{a list that defines a partition of an vector on the form \code{0:(q-1)},
e.q. the vector that enumerates the \code{q} joint outcomes of}

\item{lens}{(integer vector) number of outcomes in each part of \code{P}}
}
\value{
an vector of length \code{unlist(P)} that assigns each outcome two a part in the partition.
}
\description{
Unlist a partition
}
\examples{

# list all configurations of two categorical variables
levels <- list(0:1, 0:2)
conf <- expand.grid(levels)

# enumerate joint outcomes
nlev <- lengths(levels)
stride <- c(1, cumprod(nlev[-length(nlev)]))
joint <- as.matrix(conf)\%*\%stride

# define partition of the outcome space
P <- list(c(0, 1, 3, 5), 2, 4)

cbind(conf, joint, get_parts(P))
}
