% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ida_cat.R
\name{ida_cat}
\alias{ida_cat}
\title{Title}
\usage{
ida_cat(
  cpdag,
  data,
  adjset = c("pa", "o"),
  params,
  x = 1:ncol(data),
  y = 1:ncol(data),
  contrasts = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cpdag}{(integer matrix) an adjacency matrix representing a CPDAG, where
\itemize{
\item \code{cpdag[i, j] = 1} indicates an edge from node \code{i} to node \code{j},
\item \code{cpdag[i, j] = 1} indicates absence of an edge from node \code{i} to node \code{j},
\item \code{cpdag[i, j] = cpdag[j, i] = 1} indicates an undirected edge from node \code{i} to node \code{j}.
}}

\item{data}{(integer matrix) data matrix representing categorical variables. Assumed to be zero-based.}

\item{adjset}{(character) which adjustment set to use in backdoor formula.}

\item{params}{(list) list with hyperparameters for the local bdeu-priors.}

\item{x, y}{(integer vector) position of cause and effect variables, respectively.}

\item{verbose}{(logical) print output to follow progression.}
}
\value{
a bida-object.
}
\description{
Title
}
\examples{

data(bida_example_cat)
attach(bida_example_cat)

# infer a CPDAG with pcalg::pc
suffStat <- list(dm = data,
                 nlev = apply(data, 2, max)+1,
                 adaptDF = FALSE)
fit <- pcalg::pc(suffStat,
                 pcalg::disCItest,
                 alpha = .05,
                 labels = colnames(data),
                 u2pd = c("retry"),
                 verbose = FALSE)
cpdag <- as(fit@graph, "matrix")
# Rgraphviz::plot(fit@graph)
cpdag

# estimate intervention probabilities
params <- list(ess = 1, nlev = apply(data, 2, max)+1)
fit <- ida_cat(cpdag, data, "pa", params, contrasts = list(jsd = jsd))
posterior_mean()

# optimal-IDA, set effects to zero exactly
res <- ida_cat(cpdag, data, "o", params, contrasts = list(jsd = jsd))
res
}
