% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv_probs.R
\name{interv_probs}
\alias{interv_probs}
\title{Compute discrete intervention distributions}
\usage{
interv_probs(obj, method = "bnlearn")
}
\arguments{
\item{obj}{a object of class \code{bnlearn::bn.fit.dnet} of length \code{n} or
a list of CPTs stored as arrays.}

\item{method}{either \code{"bnlearn"} (default) or \code{"exact"}.}
}
\value{
a \code{n}-by-\code{n} matrix with intervention probabilities
}
\description{
For each variable in a discrete Bayesian network and every possible intervention
on that variable, compute the marginal intervention distribution of all other
variables in the network.
The default version is based on \link[bnlearn:cpquery]{bnlearn::mutilated} and \link[bnlearn:cpquery]{bnlearn::cpdist},
and estimates the interventional distributions by sampling from the network
under all possible interventions.
An exact version is via \link[SumProductVE:cpquery]{SumProductVE::cpquery} is applicable to smaller networks.
}
\examples{
bn <- readRDS("./data/alarm.rds")

pdo <- interv_probs(bn)
pdo

system.time(exact <- interv_probs(bn, method = "exact"))
system.time(pdo2  <- interv_probs(bn))

plot(unlist(pdo), unlist(exact))
abline(a = 0, b = 1, col = "red")

plot(unlist(pdo), unlist(pdo2))
abline(a = 0, b = 1, col = "red")


}
