% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv_probs.R
\name{interv_probs_from_bn}
\alias{interv_probs_from_bn}
\title{Approximate intervention probabilities}
\usage{
interv_probs_from_bn(bn, x, y)
}
\arguments{
\item{bn}{an object of class bn.fit.dnet}
}
\value{
a list of length \code{length(y)} with the interventional probability tables
}
\description{
Approximate intervention probabilties under interventions on single\code{x},
using \link[bnlearn:cpquery]{bnlearn::mutilated} and \link[bnlearn:cpquery]{bnlearn::cpdist}.
}
\keyword{internal}
