% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv_probs.R
\name{interv_probs_from_cpts}
\alias{interv_probs_from_cpts}
\title{Compute exact intervention probabilities}
\usage{
interv_probs_from_cpts(cpts, x, y, dmat)
}
\arguments{
\item{cpts}{a list of of length \code{n} with CPTs stored in arrays, where \code{names(dimnames(cpts[[i]]))} gives
the scope of the \code{i}th CPT. Note that}
}
\value{
a list of length \code{length(y)} with the interventional probability tables
}
\description{
Compute exact intervention probabilities under interventions on single cause variable \code{x}
}
\keyword{internal}
