% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv_probs.R
\name{interv_probs_x_bn}
\alias{interv_probs_x_bn}
\title{Compute approximate intervention probabilities under interventions on \code{x},
using \link[bnlearn:cpquery]{bnlearn::mutilated} and \link[bnlearn:cpquery]{bnlearn::cpdist}.}
\usage{
interv_probs_x_bn(bn, x, y, dmat)
}
\arguments{
\item{bn}{a object of class \code{bnlearn::bn.fit.dnet} of length \code{n}}
}
\value{
a list of \code{length(y)} with interventional cpts
}
\description{
Compute approximate intervention probabilities under interventions on \code{x},
using \link[bnlearn:cpquery]{bnlearn::mutilated} and \link[bnlearn:cpquery]{bnlearn::cpdist}.
}
