% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsd.R
\name{jsd}
\alias{jsd}
\title{Compute the JSD of a conditional probability table}
\usage{
jsd(p, digits = 10)
}
\arguments{
\item{p}{(numeric array) a CPT stored in a \code{r-by-q} matrix or a collection of
\code{n} such CPTs stored in a \code{r-by-q-by-n} array, where \code{r} is the cardinality of
the outcome and \code{q} the cardinality of the parent variable(s).}

\item{digits}{(numeric) precision of output.}
}
\value{
a numeric array with the JSD of each CPT in \code{p}
}
\description{
Compute the Jenson-Shannon-Divergence (JSD) of the categorical distributions
described by a conditional probability table (CPT).
}
\examples{

a <- .25
p1 <- matrix(c(a, 1-a, 1-a, a), nrow = 2, ncol = 2)
all(colSums(p1) == 1)  # sum over outcomes equals 1
jsd(p1)

p2 <- matrix(c(a, 1-a, a, 1-a), nrow = 2, ncol = 2)
jsd(p2) # the JSD beteen identical distributions is 0

p3 <- diag(2)
jsd(p3) == log(2) # log(q) is the upper bound
jsd(p3)

p3 <- array(c(p1, p2, p3), c(2, 2, 3))
jsd(p3) # compute the JSD of a collection of CPTs stored in a array

}
