% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor_sample_cat.R
\name{mean_bdeu}
\alias{mean_bdeu}
\title{Title}
\usage{
mean_bdeu(Nyx, ess = 1, partition = NULL)
}
\arguments{
\item{partition}{}
}
\description{
Title
}
\examples{


Nyxz <- bida_sparse_array(1:3, 0:2, c(3, 3, 3))
partition <- list(0:2, 3:27)
}
