% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_famscore.R
\name{optimize_famscore_bdeu}
\alias{optimize_famscore_bdeu}
\title{Optimize local bdeu score}
\usage{
optimize_famscore_bdeu(
  data,
  levels,
  nlev = lengths(levels),
  j,
  parentnodes,
  ess = 1,
  method = NULL,
  ...
)
}
\arguments{
\item{j}{(integer)
column index of node}

\item{parentnodes}{(integer vector)
column indicies of parent nodes. A vector of length 0 (e.g. \code{integer(0)})
translates to no parents.}

\item{ess}{(numeric constant)
equivalent sample size.}

\item{method}{(character)
algorithm to learn local structure. See \link{optimize_partition}.
If \code{NULL} (default) no local structure is learned. Ignored if \code{length(parentnodes) < 2}.}

\item{...}{additional parameters sent to \code{\link[=optimize_partition]{optimize_partition()}}.}
}
\value{
a numeric constant, the Bdeu-score. If a local structure is inferred,
this is returned as an named attribute (\code{structure}).
}
\description{
Optimize local bdeu score
}
