% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_partition_from_data.R
\name{optimize_partition_from_data}
\alias{optimize_partition_from_data}
\alias{optimize_partition_from_df_pcart}
\title{Title}
\usage{
optimize_partition_from_data(
  data,
  j,
  parentnodes,
  ess,
  nlev,
  method,
  verbose = FALSE,
  ...
)

optimize_partition_from_df_pcart(
  df,
  j,
  parentnodes,
  ess,
  nlev,
  verbose = FALSE
)
}
\arguments{
\item{ess}{}

\item{df}{a data.frame of factor variables, with levels corresponding to nlev.}
}
\description{
Title
}
\examples{

# binary split data with missing levels
data <- cbind(z = rep(0:2, 2),
              x = rep(c(0, 2), each = 3),
              y = c(rep(0, 3), rep(1, 3)))
nlev <- rep(3, 3)
pcart <- optimize_partition_from_data(data, 3, 1:2, 1, nlev, "pcart", verbose = TRUE)
tree  <- optimize_partition_from_data(data, 3, 1:2, 1, nlev, "tree", regular = FALSE, verbose = TRUE)
ptree <- optimize_partition_from_data(data, 3, 1:2, 1, nlev, "ptree", verbose = TRUE)

counts <- matrix(table(data.frame(data)), 6, 2)
}
