% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_partition_from_data.R
\name{optimize_partition_from_df_pcart}
\alias{optimize_partition_from_df_pcart}
\alias{get_rules.pcart}
\title{Greedy optimization of local structure using \code{rpcart}}
\usage{
optimize_partition_from_df_pcart(
  df,
  ess,
  verbose = FALSE,
  use_structure_score = FALSE
)

\method{get_rules}{pcart}(obj)
}
\arguments{
\item{df}{(data frame) an data frame with factors.
The first column includes the repsonse variable, the remaining predictor variables.}
}
\value{
an object of class \code{partition} with sub-class \code{pcart}
}
\description{
This function learn a partition of the parent space of an variable,
defined by the binary decision tree that maximizes the score.
It is a wrapper around \code{rpcart:::opt.pcart.cat.bdeu} that fits the optimal binary tree.
If the variables in \code{df} are not factor variables, they will be converted to factors
with levels corresponding to those observed. Note that the score will depend
on wheter or not missing levels are included or not.
}
\seealso{
\code{\link[=optimize_partition_from_data]{optimize_partition_from_data()}}
}
\keyword{internal}
