% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_partition_tree_sparse.R
\name{optimize_partition_tree_sparse}
\alias{optimize_partition_tree_sparse}
\title{Title}
\usage{
optimize_partition_tree_sparse(
  counts,
  ess,
  min_score_improv = 0,
  prune = F,
  simplify = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{coord}{}
}
\description{
Title
}
\examples{

dnames <- list(y = 0:1, x = 0:1, z = 0:1)
counts <- bida_sparse_array(c(10, 100, 10, 100, 10, 1000, 10, 1000),
                            index = 0:7,
                            dim = lengths(dnames),
                            dimnames = dnames)
ess <- 1
min_score_improv <- -Inf
prune <- TRUE
tree <- optimize_partition_tree_sparse(counts, ess, min_score_improv, prune, verbose = TRUE)

summary.tree(tree)
predict.tree(tree, expand.grid(dnames))
}
