% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_from_labels.R
\name{partition_from_labels}
\alias{partition_from_labels}
\title{Convert label-set to partition}
\usage{
partition_from_labels(
  labels,
  nlev,
  levels = lapply(nlev - 1, seq.int, from = 0)
)
}
\arguments{
\item{labels}{a list with labels on each edge from each parent $X_i$ into $X_j$.
Empty labels have to be included as \code{NULL} elements in the list, as the variables are referred to by
their position not names.}

\item{nlev}{(integer vector) the cardinality of the parent variables.}

\item{levels}{(optional) a list with integer vectors, zero-indexed levels of each variable.
Included as argument for saving computation time if pre-computed. Otherwise, it can be derived from \code{nlev}.}
}
\value{
a integer vector of length \code{prod(nlev)} assigning the parent outcomes to a partition.
}
\description{
Convert the set of labels associated with a node $X_j$ in a labeled DAG to a
partition of the outcome space of $Pa(X_j)$.
}
\examples{
nlev   <- rep(2, 3)
levels <- lapply(nlev-1, seq.int, from = 0)
labels <- vector("list", 2)
labels[[1]] <- rbind(c(0, 0), c(0, 1))
cbind(expand.grid(levels), labels_to_parts(labels, nlev, levels))

labels[[2]] <- rbind(c(0, 0))  # add label to edge from parent 2
cbind(expand.grid(levels), labels_to_parts(labels, nlev, levels))
}
