% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_mean.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Compute posterior mean}
\usage{
posterior_mean(x, ...)
}
\arguments{
\item{x}{object of class bida, bida_params, or backdoor_params}

\item{...}{additional arguments}
}
\value{
the posterior mean(s) of the distribution(s) specified by \code{x}
}
\description{
Compute posterior mean
}
