% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sample.R
\name{posterior_sample.bida_pair_bdeu}
\alias{posterior_sample.bida_pair_bdeu}
\title{\itemize{
\item \code{posterior_sample}: if \code{ace_funs == NULL} a three dimensional array with IPTs.
Otherwise a matrix with the
}}
\usage{
\method{posterior_sample}{bida_pair_bdeu}(x, n, contrasts = NULL)
}
\description{
\itemize{
\item \code{posterior_sample}: if \code{ace_funs == NULL} a three dimensional array with IPTs.
Otherwise a matrix with the
}
}
