% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_sample.R
\name{posterior_sample}
\alias{posterior_sample}
\title{Sample from posterior distribution specified by bida-objects}
\usage{
posterior_sample(x, n, ...)
}
\arguments{
\item{x}{object of class bida, bida_params, or backdoor_params}

\item{n}{(integer) samplesize}

\item{...}{additional arguments}
}
\value{
a posterior sample from the distribution(s) specified by \code{x}
}
\description{
Sample from posterior distribution specified by bida-objects
}
