% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_posterior.R
\name{print.posterior_bida}
\alias{print.posterior_bida}
\title{Methods for objects of class \code{posterior_bida}}
\usage{
\method{print}{posterior_bida}(obj)
}
\description{
Methods for objects of class \code{posterior_bida}
}
