% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_cpt.R
\name{rand_cpt}
\alias{rand_cpt}
\title{Draw a conditional probability table (CPT)}
\usage{
rand_cpt(dim, scope = NULL, alpha = 1, local_structure = "none", ...)
}
\arguments{
\item{dim}{(integer vector) of cardinality of the outcome variable and the
parent variables.}

\item{scope}{(character vector) names of the variables.}

\item{alpha}{(integer)}

\item{local_structure}{(character) name of algorithm to form a partitioning /
produce parameter restrictions.
Defaults to \code{"none"}, which returns a CPT with no parameter restrictions.}

\item{...}{additional arguments sendt to \link{rand_partition}.}
}
\value{
an array with dimension \code{nlev} and dimension named by \code{scope}.
}
\description{
Construct a random conditional probability table (CPT) by sampling independently
a vector of probabilities from a Dirichlet distribution for each parent configuration.
The function returnes an array with named dimension, in the same format
as \link[bnlearn:bn.fit.class]{bnlearn::bn.fit.dnode} store CPTs.
}
\examples{

# no parents
dim <- c(2)
scope <- c("Y")
rand_cpt(dim, scope)

# two parents
dim <- c(2, 3, 4)
scope <- c("Y", "X", "Z")
rand_cpt(dim, scope)

}
\keyword{internal}
