% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_dag.R
\name{rand_dag}
\alias{rand_dag}
\title{Draw a random adjacency matrix}
\usage{
rand_dag(n, d, names = paste0("X", seq_len(n)), weighted = FALSE)
}
\arguments{
\item{n}{(integer) number of nodes}

\item{d}{(integer) expected number of neighbours}

\item{weighted}{(boolean) wheter each edge should be assigned weights. Default is FALSE.}
}
\value{
a n-by-n adjacency matrix
}
\description{
Wrapper around \link[pcalg:randDAG]{pcalg::randDAG} returning a adjacency matrix rather than a graphNEL object
}
