% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_dist.R
\name{rand_dist}
\alias{rand_dist}
\title{Draw a random distribution over a DAG}
\usage{
rand_dist(dag, type = "cat", ...)
}
\arguments{
\item{dag}{(integer matrix) an adjacency matrix}

\item{type}{(character) type of distribution}

\item{...}{additional arguments for the function producing each local distribution.
See \code{\link[=rand_cpt]{rand_cpt()}} for categorical distributions.}
}
\value{
a list with distributions
}
\description{
Draw a random distribution over a DAG.
The distribution is represented in a format matching \link[bnlearn:bn.fit]{bnlearn::bn.fit} class,
and is used to construct such objects.
}
\examples{

dag  <- matrix(0, 3, 3, dimnames = list(c("Z", "X", "Y"), c("Z", "X", "Y")))
dag[upper.tri(dag)] <- 1
rand_dist(dag, "cat", nlev = rep(3, ncol(dag)))

}
\seealso{
\code{\link[=rand_cpt]{rand_cpt()}}
}
