% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_partition.R
\name{rand_partition}
\alias{rand_partition}
\title{Title}
\usage{
rand_partition(nlev, prob, method = "labels")
}
\description{
Title
}
\examples{

set.seed(007)
nlev <- c(2, 2, 2)
rand_partition(nlev, prob = .5)

}
