% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_fill}
\alias{rbind_fill}
\title{Combine vectors of various length}
\usage{
rbind_fill(x)
}
\arguments{
\item{x}{(list)
a list with numeric vectors}
}
\value{
a \code{n-by-m} with \code{n = min(1, length(x))} and \code{m = min(1, max(lengths(x)))}.
}
\description{
This function combines a list of vectors into a matrix (row-wise),
padding vectors of less-than-maximal length with NA values.
Used to store adjustment sets in a matrix in \link{adjsets_support_from_dags}.
}
\keyword{internal}
