% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowsum_fast.R
\name{rowsum_fast}
\alias{rowsum_fast}
\alias{rowsum_fast.default}
\alias{rowsum_fast.matrix}
\alias{rowsum_fast.bida_sparse_array}
\title{Compute sums over grouping variable}
\usage{
rowsum_fast(x, group, ugroup, na.rm = FALSE)

\method{rowsum_fast}{default}(x, group, ugroup = unique(group), na.rm = FALSE)

\method{rowsum_fast}{matrix}(x, group, ugroup = unique(group), na.rm = FALSE)

\method{rowsum_fast}{bida_sparse_array}(x, group, ugroup, na.rm = TRUE)
}
\arguments{
\item{x}{(numeric) vector or matrix}

\item{group}{grouping variable}

\item{ugroup}{unique elements of grouping variable. Also defines the order of
the group sums returned by the function.}

\item{na.rm}{See \link[base:rowsum]{base::rowsum}.}
}
\value{
An array or a matrix, depending on the class of x.
}
\description{
Speedy version of \link[base:rowsum]{base::rowsum} that assumes takes the unique elements of
grouping variable as an argument, and calls the internal method without any
error check.
}
\examples{

# grouped sum of vector
x <- rep(1:3, 3)
ugroup <- letters[3:1]
group <- rep(ugroup, 3)
rowsum(x, group, reorder = F)
rowsum_fast(x, group, ugroup)

microbenchmark::microbenchmark(rowsum(x, group, reorder = F), rowsum_fast(x, group, ugroup))

# grouped sum of matrix
x <- matrix(rep(1:9, 3), nrow = 9, ncol = 3)
rowsum(x, group, reorder = F)
rowsum_fast(x, group, ugroup)

microbenchmark::microbenchmark(rowsum(x, group, reorder = F), rowsum_fast(x, group, ugroup))

#' # no error on misspecification
rowsum_fast(x, group[-1], ugroup, na.rm = F)


# with NAs
x <- c(rep(1:3, 3), NA)
ugroup <- letters[3:1]
group <- c(rep(ugroup, 3), letters[3])
rowsum_fast(x, group, ugroup, na.rm = F)
rowsum_fast(x, group, ugroup, na.rm = T)

}
