% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_from_bn.R
\name{sample_data_from_bn}
\alias{sample_data_from_bn}
\title{Sample data from bayesian networks}
\usage{
sample_data_from_bn(bn, samplesize, maxiter = 10^4)
}
\arguments{
\item{bn}{an object of class bn.fit.dnet}

\item{samplesize}{sample size}

\item{maxiter}{(integer) maximum numer of iterations}
}
\value{
a data frame with factor variables
}
\description{
Wrapper around function \link{bnlearn::rbn},
that resamples data sets until all variables has at least 2 observed levels.
}
