% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_from_cpts.R
\name{sample_from_cpt}
\alias{sample_from_cpt}
\title{Draw a categorical variable given its parents}
\usage{
sample_from_cpt(cpt, parents)
}
\arguments{
\item{cpt}{(numeric array) a CPT}

\item{parents}{(integer vector) parent configuration of each sample}
}
\value{
an array of length \code{NROW(parents)} sampled from the CPT
}
\description{
Draw a categorical variable given its parents
}
\examples{

nlev <- c(3, 3, 3)
r <- nlev[1]        # cardinality of outcome variable
q <- prod(nlev[-1]) # cardinality of parent variables
cpt <- t(rDirichlet(q, rep(1, r)))
dim(cpt) <- nlev

# sample parents uniformily
parents <- sample.int(q, 10**4, TRUE)

# sample from cpt
y <- sample_from_cpt(cpt, parents)

# compare means
tabulate(y + 1, r)/length(y)
rowMeans(cpt)
}
\keyword{internal}
