% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dag.R
\name{score_dag}
\alias{score_dag}
\title{Compute Bayesian score of DAG}
\usage{
score_dag(data, dag, type = "cat", params)
}
\arguments{
\item{data}{(matrix) data matrix}

\item{type}{(character) name of score}

\item{params}{(list) list with additional parameters}

\item{j}{(integer) column position of node}

\item{parentnodes}{(integer vector) column position of parents of node \code{j}}

\item{ess}{(integer) equivalent sample size}

\item{nlev}{(integer vector) cardinality of each variable}

\item{partitions}{(list) list with partitioning of the parent space of node \code{j}}
}
\value{
(numeric constant) the score of node \code{j} given \code{parentnodes}
}
\description{
Compute Bayesian score of DAG
}
\examples{

n   <- 3
dag <- matrix(0, n, n)
dag[upper.tri(dag)] <- 1


# Categorical data ---
params <- list(nlev = 2:4, ess = 1)

## draw random data set
N <- 100
data <- sapply(params$nlev, sample, size = N, replace = T) -1

## test score equivalence
score1 <- score_dag(data, dag, type = "cat", params)
score2 <- score_dag(data, t(dag), type = "cat", params)
stopifnot(score1 == score2)


# labeled DAG - add list of partitions of each CPT
params$partition <- list(NULL, NULL, list(c(0, 1), 2, 3))
score_dag(data, dag, type = "cat", params)


}
