% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_and_write_to_file}
\alias{sim_and_write_to_file}
\title{Run simulation and write result to file}
\usage{
sim_and_write_to_file(dir_out, filename, run, ...)
}
\arguments{
\item{dir_out}{(character) path to directory where results should be stored}

\item{filename}{(character) name of path}

\item{run}{(function) a function that runs a simulation}

\item{...}{additional arguments, sent to \code{run}.}

\item{id}{(character) id of simulation}
}
\value{
If \verb{is.null(dir_out):} the output of \code{run}. Otherwise the output is written to
file and the function returns \code{NULL}.
}
\description{
Wrapper function around some simulation \code{run} that only run if the results are
not already present in the directory defined by \code{dir_out}
}
