% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_plot.R
\name{sim_run}
\alias{sim_run}
\title{Title}
\usage{
sim_run(N, r, partname, n, k, verbose = F)
}
\arguments{
\item{N}{(integer) sample size}

\item{r}{(integer) iteration number}

\item{k}{(integer) cardinality of outcome}

\item{verbose}{(bolean)}

\item{parts}{(integer vector)}

\item{levels}{(list) levels of parent variables}
}
\value{
matrix
}
\description{
Title
}
\examples{

levels <- list(0:1, 0:1)
parts  <- c(1, 1, 3, 4)
sim_run(N, r, parts, levels, k = 2, verbose = TRUE)
}
