% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_out.R
\name{sum_out}
\alias{sum_out}
\title{Compute marginal sums}
\usage{
sum_out(x, MARGIN)
}
\arguments{
\item{x}{an object of class \code{base::array} or \code{bida_sparse_array}}

\item{MARGIN}{(integer vector)}
}
\value{
an array with marginal sums
}
\description{
Compute marginal sums
}
\examples{

arr <- array(1:27, c(3, 3, 3), list(x = 0:2, y = 0:2, z = 0:2))
sum_out(arr, 2)
sum_out(arr, 2:3) == colSums(arr, dims = 1)
sum_out(arr, 1:2) == rowSums(arr, dims = 2)
sum_out(arr, c(1, 3)) == apply(arr, c(1, 3), sum)

microbenchmark::microbenchmark(sum_out(arr, c(1, 3)), apply(arr, c(1, 3), sum))

# sparse array
sarr <- as.bida_sparse_array(arr)
}
