% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjsets_support_from_cpdag.R
\name{adjsets_support_from_cpdag}
\alias{adjsets_support_from_cpdag}
\title{List unique adjustment sets in a CPDAG}
\usage{
adjsets_support_from_cpdag(adjsets, cpdag)
}
\arguments{
\item{cpdag}{(matrix)}
}
\description{
List unique adjustment sets in a CPDAG
}
\details{
\itemize{
\item \code{adjsets_support_from_cpdags}:
Identifies unique adjustment sets (parent or o-set only) locally in a CPDAG,
by directing siblings of every node. The same set of variables can consititute
the o-set in multiple of the resulting (P)DAGs, reflected in the returned "support".
}
}
