% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdoor_params.R
\name{backdoor_params}
\alias{backdoor_params}
\title{Compute parameters for direct backdoor estimation of an intervention distribution}
\usage{
backdoor_params(type, data, x, y, z, hyperpar, lookup)
}
\arguments{
\item{data}{a N-by-n data matrix}

\item{x}{(integer)
column position of cause variable}

\item{y}{(integer)
column position of effect variable}

\item{z}{(integer vector) column position of the adjustment variables.
If \code{any(y == z)}, the marginal distribution of \code{y} is returned.
If \code{length(z) = 0}, the conditional distribution of \code{y} given \code{x}.}

\item{hyperpar}{list of parameters, depending on the prior distribution. See details.}
}
\value{
\itemize{
\item If \code{type} is in \code{c("cat", "ldag", "tree")}, an object of class \link{bida_bdeu}.
}
}
\description{
Compute parameters for direct backdoor estimation of an intervention distribution
}
\examples{

# Categorical data ----
nlev <- 2:4
lev  <- lapply(nlev-1, seq.int, from = 0)
data <- as.matrix(expand.grid(lev))


y <- 3
x <- 2
z <- 1
hyperpar <- list(nlev = nlev, ess = 1)

# no adjustment set, return params for cond distribution p(y|x)
backdoor_params("cat", data, x, y, integer(0), hyperpar, NULL)

# y is in adjustment set, return params for marginal distribution
backdoor_params("cat", data, x, y, c(y, z), hyperpar, NULL)

# with optimized partitioning
type <- "tree"
bdeu_tree <- backdoor_params(type, data, x, y, z, hyperpar, NULL)

# compare with results from lookup table after running BiDAG::sampleBN with user-score
lookup   <- rlang:::new_environment()
hyperpar$edgepf <- 1
scorepar <- define_scoreparameters(data, "tree", hyperpar, lookup = lookup)
BiDAG:::usrDAGcorescore(y, c(x, z), n = 3, scorepar)
ls.str(lookup)
bdeu_lookup <- backdoor_params(type, data, y, x, z, hyperpar, lookup)
stopifnot(all.equal(bdeu_tree, bdeu_lookup))

}
