% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_.R
\name{bida}
\alias{bida}
\alias{posterior_mean.bida}
\alias{posterior_sample.bida}
\title{Class: bida}
\usage{
bida(
  type = "categorical",
  dags,
  data,
  adjset = "pa",
  hyperpar,
  x = 1:ncol(data),
  y = 1:ncol(data)
)

\method{posterior_mean}{bida}(x, ...)

\method{posterior_sample}{bida}(x, n, ...)
}
\arguments{
\item{data}{(matrix)}

\item{x}{an object of class \link{bida}}

\item{ps}{(list)
Support over adjustment sets, stored in a list with two named elements \code{sets} and \code{support}.
See \link{adjsets_support_from_dags} \link{parent_support_from_dags} \link{parent_support_exact}}

\item{params}{additional params}

\item{pairs}{(matrix) a n-by-n matrix indicating which pairs should be considered.
If NULL (default), all pairs are considered and \code{pairs[i, j] = 1} for all but \code{i == j}.}

\item{samplesize}{(integer) samplesize}
}
\value{
\code{bida}: a n-by-n list where each element is an object of class \code{bida_pair}
}
\description{
An object of class bida is a matrix list with objects of class \link{bida_pair},
that describes the posterior mixture distribution over the causal effect between every variable pair.
The \code{posterior_mean} and \code{posterior_sample} methods calls the respective methods
for each \link{bida_pair} and collect the output in a matrix list.
}
\examples{
dag  <- rbind(c(0, 1, 1), c(0, 0, 1), c(0, 0, 0))
colnames(dag) <- rownames(dag) <- letters[1:3]

# Categorical data
nlev <- 2:4
bn <- rand_bn(dag, "cat", alpha = 1, nlev = nlev)
data <- sample_data_from_bn(bn, 10**4)

# list parent sets in true DAG
ps <- parent_support_from_dags(list(dag))

params <- list(nlev = nlev, ess = 1)
fit <- bida(ps, data, "cat", params)
means <- posterior_mean(fit)
means

# compare with true distributions
pdo <- interv_probs_from_bn(bn, "exact")
dindx <- diag(ncol(dag)) == 1
plot(unlist(pdo[!dindx]), unlist(means[!dindx]))

}
\seealso{
\link{bida_pair_cat}
}
