% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida.R
\name{bida}
\alias{bida}
\alias{posterior_mean.bida}
\alias{posterior_sample.bida}
\title{Class: bida}
\usage{
bida(
  dags,
  data,
  type = "cat",
  adjset = "pa",
  params,
  x = 1:ncol(data),
  y = 1:ncol(data),
  replace_large_adjset = NULL,
  verbose = FALSE
)

\method{posterior_mean}{bida}(obj, ...)

\method{posterior_sample}{bida}(obj, n, ...)
}
\arguments{
\item{data}{(matrix)}

\item{params}{additional params}

\item{x}{an object of class \link{bida}}

\item{ps}{(list)
Support over adjustment sets, stored in a list with two named elements \code{sets} and \code{support}.
See \link{adjsets_posterior}}

\item{pairs}{(matrix) a n-by-n matrix indicating which pairs should be considered.
If NULL (default), all pairs are considered and \code{pairs[i, j] = 1} for all but \code{i == j}.}

\item{samplesize}{(integer) samplesize}
}
\value{
\code{bida}: a n-by-n list where each element is an object of class \code{bida_pair}
}
\description{
An object of class bida is a matrix list with objects of class \link{bida_posterior},
that describes the posterior mixture distribution over the causal effect between every variable pair.
The \code{posterior_mean} and \code{posterior_sample} methods calls the respective methods
for each \link{bida_posterior} and collect the output in a matrix list.
}
\examples{

dag  <- rbind(c(0, 1, 1), c(0, 0, 1), c(0, 0, 0))
colnames(dag) <- rownames(dag) <- letters[1:3]
dags <- list(dag, t(dag))

# Categorical data
nlev <- 2:4
cpts <- rand_dist(dag, "cat", nlev = nlev)
bn   <- custom_bn(dag, cpts)
data <- sample_data_from_bn(bn, 10**4)
head(data)


params <- list(nlev = nlev, ess = 1)
bida <- bida(list(dag, t(dag)), data, "cat", "pa", params)
bida

# posterior mean
means <- posterior_mean(bida)
means

# posterior mean of intervention effect (jsd)
means <- posterior_mean(bida, contrasts = list(jsd = jsd))
means
}
\seealso{
\link{bida_pair_cat}
}
