% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_sparse_array.R
\name{bida_sparse_array}
\alias{bida_sparse_array}
\alias{new_bida_sparse_array}
\title{class: cida sparse array}
\usage{
new_bida_sparse_array(x, index, dims, names = NULL)
}
\arguments{
\item{x}{(numeric vector) values of non-zero elements}

\item{index}{(integer vector) position of non-zero elements}

\item{dims}{(integer vector) dimension of array}
}
\value{
an object of class \code{bida_sparse_array}
}
\description{
Sparse representation of multi-dimensional arrays.
}
\examples{
x <- array(c(0, 0, 1, 0, 0, 1, 0, 0, 1), dim = c(3, 3))
y <- new_bida_sparse_array(c(1, 1, 1), c(3, 6, 9), c(3, 3))
y
all.equal(x, as.array(x))
}
\keyword{internal}
