% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bida_pair.R, R/posterior_sample.R
\name{posterior_sample.bida_pair}
\alias{posterior_sample.bida_pair}
\alias{posterior_sample}
\title{Sample from posterior distribution specified by bida-objects}
\usage{
\method{posterior_sample}{bida_pair}(x, n, contrasts = NULL)

posterior_sample(x, n, contrasts, reduced)
}
\arguments{
\item{x}{object of class bida, bida_params, or backdoor_params}

\item{n}{(integer) samplesize}

\item{...}{additional arguments}
}
\value{
a posterior sample from the distribution(s) specified by \code{x}
}
\description{
Sample from posterior distribution specified by bida-objects
}
\keyword{internral}
