% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_bn.R
\name{rand_bn}
\alias{rand_bn}
\title{Draw a random Bayesian network}
\usage{
rand_bn(n, d, type = "cat", ..., attr_to_keep = character(0))
}
\arguments{
\item{type}{(character) type of distribution.}

\item{...}{}

\item{dag}{(integer matrix) adjacency matrix.}
}
\value{
a \code{bnlearn::bn.fit} object
}
\description{
Draw a random DAG and a distribution over the DAG, and store the BN as a \code{bnlearn::bn.fit} object.
}
\examples{

n <- 3
dag <- matrix(0, n, n)
dag[upper.tri(dag)] <- 1


# categorical ---
# draw random DAG
set.seed(007)
dag <- rand_dag(3, 2)
dag

# draw random distrib over DAG
cpts <- rand_dist(dag, "cat", nlev = nlev)

# create bn object
bn <- custom_bn(dag, cpts)

# replicate with rand_bn()
set.seed(007)
bn2 <- rand_bn(3, 2, "cat", nlev = nlev)
stopifnot(all.equal(bn, bn2))

}
\keyword{internal}
