% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_bn.R
\name{rand_bn}
\alias{rand_bn}
\title{Title}
\usage{
rand_bn(dag, type = "cat", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

n <- 3
dag <- matrix(0, n, n)
dag[upper.tri(dag)] <- 1

# categorical distribution with local structure
nlev <- rep(2, ncol(dag))
partitions <- list(NULL, NULL, X3 = list(0:2, 1))
rand_bn(dag, "cat", alpha = 1, nlev = rep(2, 3), partitions)
}
\keyword{internal}
