% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_bn.R
\name{rand_partition}
\alias{rand_partition}
\title{Draw a random partition of an outcome space}
\usage{
rand_partition(nlev, prob, method = "tree", regular = FALSE, ...)
}
\arguments{
\item{nlev}{cardinality of the variables.}

\item{prob}{parameter controlling the size of the partition.}

\item{method}{name of algorithm.}

\item{regular}{If \code{TRUE}, the partition is forced to be regular. See \link{make_regular}.}
}
\value{
an vector of length \code{prod(nlev)} assigning each joint outcome to a
subset of the partition.
}
\description{
Draw a random partition of the outcome space of a set of categorical variables
}
\examples{

set.seed(1)
nlev <- c(2, 2, 2)
rand_partition(nlev, .75, method = "tree")

# tree of given depth
rand_partition(nlev, 1, "tree", maxdepth = 1)
rand_partition(nlev, 1, "tree", maxdepth = .5)

# tree of given depth, forced to be regular
rand_partition(nlev, 1, "tree", maxdepth = 1, regular = TRUE)
}
