% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowsum_fast}
\alias{rowsum_fast}
\title{Compute sum conditional on a grouping variable}
\usage{
rowsum_fast(x, group, ugroup, na.rm = FALSE)
}
\arguments{
\item{x}{a matrix, data frame or vector of numeric data.  Missing
    values are allowed.  A numeric vector will be treated as a column vector.}

\item{group}{a vector or factor giving the grouping, with one element
    per row of \code{x}.  Missing values will be treated as another
    group and a warning will be given.}

\item{ugroup}{unique groups}

\item{na.rm}{logical (\code{TRUE} or \code{FALSE}).  Should \code{NA}
    (including \code{NaN}) values be discarded?}
}
\description{
Faster version of \link[base:rowsum]{base::rowsum}, to be used when the unique levels of the grouping variable is known.
Note that the sanity-checks of rowsum is excluded.
}
\keyword{internal}
